+++ example/localization/localization.zodart.dart
@@ -121,7 +121,12 @@ final class _PersonSchemaUtils
 
   @override
   ZObject<Person> get zObject {
-    return ZObject.withMapper(_schemaMap, fromJson: _toResult);
+    return ZObject.withTypedCrossFieldValidation(
+      _schemaMap,
+      fromJson: _toResult,
+      crossValidators: <CrossFieldValidator<PersonSchemaFieldAccessor>>[],
+      parsedFieldAccessorFactory: PersonSchemaFieldAccessor.new,
+    );
   }
 
   @override
@@ -132,4 +137,15 @@ final class _PersonSchemaUtils
     firstName: val['firstName'] as String,
     lastName: val['lastName'] as String,
   );
+}
+
+/// Type-safe parsed fields accessor for [PersonSchema].
+class PersonSchemaFieldAccessor extends ParsedFieldAccessor {
+  PersonSchemaFieldAccessor(super.schema, super.parsedValues);
+
+  int? get age => (this['age'] as int?);
+
+  String get firstName => (this['firstName'] as String);
+
+  String get lastName => (this['lastName'] as String);
 }