+++ lib/src/code_generation/schema_parsing/spec_builder_input.dart
@@ -8,17 +8,21 @@ import 'schema_parsing.dart';
 /// This class defines the common shape of all spec builder inputs, including:
 /// - [refs]: A collection of shared references (types, imports, etc.)
 /// - [schema]: The parsed schema representing the shape of data to generate code for.
+/// - [crossFieldValidators]: A list of cross-field validators.
 ///
 /// Use [accept] with a [SpecBuilderInputVisitor] to dispatch behavior based on the specific subclass.
 sealed class SpecBuilderInput {
   /// Creates a new [SpecBuilderInput].
-  const SpecBuilderInput({required this.refs, required this.schema});
+  const SpecBuilderInput({required this.refs, required this.schema, required this.crossFieldValidators});
 
   /// Shared references used during code generation (e.g., for types or namespacing).
   final Refs refs;
 
   /// The parsed schema (input and output types) used for generation.
   final Schema schema;
+
+  /// Cross-field validators names list.
+  final List<String> crossFieldValidators;
 
   /// Accepts a [SpecBuilderInputVisitor] and dispatches to the appropriate subtype method.
   List<Spec> accept(SpecBuilderInputVisitor visitor);
@@ -29,7 +33,11 @@ sealed class SpecBuilderInput {
 /// Used when an annotated class requests generation of a new output class.
 class GenerateNewClassSpec extends SpecBuilderInput {
   /// Creates a new [GenerateNewClassSpec] with the given references and schema.
-  const GenerateNewClassSpec({required super.refs, required super.schema});
+  const GenerateNewClassSpec({
+    required super.refs,
+    required super.schema,
+    required super.crossFieldValidators,
+  });
 
   @override
   List<Spec> accept(SpecBuilderInputVisitor visitor) => visitor.visitGenerateNewClassSpecInput(this);
@@ -40,7 +48,12 @@ class GenerateNewClassSpec extends SpecBuilderInput {
 /// This is used when an annotation specifies an existing class.
 class UseExistingClassSpec extends SpecBuilderInput {
   /// Creates a new [UseExistingClassSpec] with the given constructor, references, and schema.
-  const UseExistingClassSpec({required this.ctor, required super.refs, required super.schema});
+  const UseExistingClassSpec({
+    required this.ctor,
+    required super.refs,
+    required super.schema,
+    required super.crossFieldValidators,
+  });
 
   /// The constructor selected from the target class that will be used to build the output object.
   final Ctor ctor;
@@ -54,7 +67,11 @@ class UseExistingClassSpec extends SpecBuilderInput {
 /// This is used when an annotation specifies a record.
 class UseRecordSpec extends SpecBuilderInput {
   /// Creates a new [UseRecordSpec] with the given references, and schema.
-  const UseRecordSpec({required super.refs, required super.schema});
+  const UseRecordSpec({
+    required super.refs,
+    required super.schema,
+    required super.crossFieldValidators,
+  });
 
   @override
   List<Spec> accept(SpecBuilderInputVisitor visitor) => visitor.visiCreateRecordSpecInput(this);