+++ test/src/code_generation/generators/all_types/goldens/zodart_existing_class.dart
@@ -163,7 +163,12 @@ final class _EUtils implements ZGenSchemaUtils<_EPropsWrapper, OutputClass> {
 
   @override
   ZObject<OutputClass> get zObject {
-    return ZObject.withMapper(_schemaMap, fromJson: _toResult);
+    return ZObject.withTypedCrossFieldValidation(
+      _schemaMap,
+      fromJson: _toResult,
+      crossValidators: <CrossFieldValidator<EFieldAccessor>>[],
+      parsedFieldAccessorFactory: EFieldAccessor.new,
+    );
   }
 
   @override
@@ -185,4 +190,37 @@ final class _EUtils implements ZGenSchemaUtils<_EPropsWrapper, OutputClass> {
     znObj: val['znObj'] as ({String id})?,
     znStr: val['znStr'] as String?,
   );
+}
+
+/// Type-safe parsed fields accessor for [E].
+class EFieldAccessor extends ParsedFieldAccessor {
+  EFieldAccessor(super.schema, super.parsedValues);
+
+  List<String> get zArrayOfStr => (this['zArrayOfStr'] as List<String>);
+
+  bool get zBool => (this['zBool'] as bool);
+
+  double get zDbl => (this['zDbl'] as double);
+
+  DateTime get zDt => (this['zDt'] as DateTime);
+
+  int get zInt => (this['zInt'] as int);
+
+  ({String id}) get zObj => (this['zObj'] as ({String id}));
+
+  String get zStr => (this['zStr'] as String);
+
+  List<String?>? get znArrayOfStr => (this['znArrayOfStr'] as List<String?>?);
+
+  bool? get znBool => (this['znBool'] as bool?);
+
+  double? get znDbl => (this['znDbl'] as double?);
+
+  DateTime? get znDt => (this['znDt'] as DateTime?);
+
+  int? get znInt => (this['znInt'] as int?);
+
+  ({String id})? get znObj => (this['znObj'] as ({String id})?);
+
+  String? get znStr => (this['znStr'] as String?);
 }